backend/addApproval/requirements.txt
azure-functions>=1.17.0
azure-data-tables>=12.4.0
python-dotenv>=1.0.0
